Rename to NavigationDrawerContainer
